+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.0.5 - 2017.12.03
+### Changed
+- Craft 3 beta 36 compatibility
+- Address Field Updates:
+    * Always show country selector
+    * Set the default country code for new entries
+
 ## 0.0.4 - 2017.06.31
 ### Added
 - Added person name field